+++ src/AtomUI.Generator/TargetMarkConstants.cs
@@ -7,5 +7,6 @@ internal static class TargetMarkConstants
     public const string ControlDesignTokenAttribute = "AtomUI.Theme.TokenSystem.ControlDesignTokenAttribute";
     public const string TokenValueConverterAttribute = "AtomUI.Theme.TokenSystem.TokenValueConverterAttribute";
     public const string GenerateDataMemberAccessorsAttribute = "AtomUI.Controls.Data.GenerateDataMemberAccessorsAttribute";
+    public const string GenerateScopedResourceHostAttribute = "AtomUI.Controls.GenerateScopedResourceHostAttribute";
     public const string LanguageSgMetaInfoAttribute = "LanguageSgMetaInfoAttribute";
 }